+++ src/main.ts
@@ -11,6 +11,9 @@ import { Engine } from "./Engine";
 import { License } from "./License";
 import { Node } from "./Node";
 import { Privileges } from "./Privileges";
+import { Proxy } from "./Proxy";
+import { SharedContent } from "./SharedContent";
+import { Scheduler } from "./Scheduler";
 import { Stream } from "./Stream";
 import { SystemRule } from "./SystemRule";
 import { ServiceCluster } from "./ServiceCluster";
@@ -36,6 +39,8 @@ import {
   IEngine,
   IExtension,
   IObject,
+  IProxyService,
+  IProxyServiceCondensed,
   IServerNodeConfiguration,
   ISystemRule,
   IHttpReturn,
@@ -43,6 +48,10 @@ import {
   ILicense,
   ILicenseAccessGroup,
   ILicenseAccessTypeCondensed,
+  ISharedContent,
+  ISharedContentCondensed,
+  ISchedulerService,
+  ISchedulerServiceCondensed,
   IServiceCluster,
   IServiceStatus,
   IStream,
@@ -51,6 +60,8 @@ import {
   ITaskExecutionResult,
   IUser,
   IUserDirectory,
+  IVirtualProxyConfig,
+  IVirtualProxyConfigCondensed,
 } from "./interfaces";
 export {
   IAbout,
@@ -67,6 +78,10 @@ export {
   IEngine,
   IExtension,
   IObject,
+  IProxyService,
+  IProxyServiceCondensed,
+  ISharedContent,
+  ISharedContentCondensed,
   IServerNodeConfiguration,
   ISystemRule,
   IHttpReturn,
@@ -74,6 +89,8 @@ export {
   ILicense,
   ILicenseAccessTypeCondensed,
   ILicenseAccessGroup,
+  ISchedulerService,
+  ISchedulerServiceCondensed,
   IServiceCluster,
   IServiceStatus,
   IStream,
@@ -82,6 +99,8 @@ export {
   ITaskExecutionResult,
   IUser,
   IUserDirectory,
+  IVirtualProxyConfig,
+  IVirtualProxyConfigCondensed,
 };
 
 import {
@@ -98,6 +117,7 @@ import {
   ITaskCreateTriggerComposite,
   ITaskCreate,
   ITableCreate,
+  ISchedulerServiceUpdate,
   IServiceClusterUpdate,
   IStreamCreate,
   ISystemRuleUpdate,
@@ -112,6 +132,10 @@ import {
   INodeCreate,
   ILicenseSetKey,
   ILicenseSetSerial,
+  IVirtualProxyUpdate,
+  ISharedContentUpdate,
+  ISharedContentCreate,
+  IProxyCreate,
 } from "./interfaces/argument.interface";
 export {
   IAppObjectUpdate,
@@ -122,6 +146,7 @@ export {
   IExtensionUpdate,
   IUserUpdate,
   IUserCreate,
+  ISchedulerServiceUpdate,
   IStreamUpdate,
   ITaskCreateTriggerSchema,
   ITaskCreateTriggerComposite,
@@ -141,6 +166,10 @@ export {
   INodeCreate,
   ILicenseSetKey,
   ILicenseSetSerial,
+  IVirtualProxyUpdate,
+  ISharedContentUpdate,
+  ISharedContentCreate,
+  IProxyCreate,
 };
 
 export class QlikRepoApi {
@@ -260,6 +289,29 @@ export class QlikRepoApi {
 
   privilegesGet = Privileges.prototype.privilegesGet;
   privilegesAssert = Privileges.prototype.privilegesAssert;
+
+  proxyAdd = Proxy.prototype.proxyAdd;
+  proxyCreate = Proxy.prototype.proxyCreate;
+  proxyGet = Proxy.prototype.proxyGet;
+  proxyGetFilter = Proxy.prototype.proxyGetFilter;
+  proxyMetadataExport = Proxy.prototype.proxyMetadataExport;
+  proxyUpdate = Proxy.prototype.proxyUpdate;
+  virtualProxyGet = Proxy.prototype.virtualProxyGet;
+  virtualProxyGetFilter = Proxy.prototype.virtualProxyGetFilter;
+  virtualProxyRemove = Proxy.prototype.virtualProxyRemove;
+  virtualProxyUpdate = Proxy.prototype.virtualProxyUpdate;
+
+  sharedContentGet = SharedContent.prototype.sharedContentGet;
+  sharedContentGetFilter = SharedContent.prototype.sharedContentGetFilter;
+  sharedContentDeleteFile = SharedContent.prototype.sharedContentDeleteFile;
+  sharedContentRemove = SharedContent.prototype.sharedContentRemove;
+  sharedContentUpdate = SharedContent.prototype.sharedContentUpdate;
+  sharedContentUploadFile = SharedContent.prototype.sharedContentUploadFile;
+  sharedContentCreate = SharedContent.prototype.sharedContentCreate;
+
+  schedulerGet = Scheduler.prototype.schedulerGet;
+  schedulerGetFilter = Scheduler.prototype.schedulerGetFilter;
+  schedulerUpdate = Scheduler.prototype.schedulerUpdate;
 
   serviceClusterCount = ServiceCluster.prototype.serviceClusterCount;
   serviceClusterGet = ServiceCluster.prototype.serviceClusterGet;